Moved Ixa error to error.rs
